Adding DRV8874 tests.
Fixed DRV8874 files to allow for testing

+++ Firmware/LCP_PistonDriver/src/encoder.h
@@ -38,14 +38,14 @@ typedef struct EncoderSettings {
     int32_t min_count;     /**< Minimum Counter setting (Default=0) */
     int32_t max_count;     /**< Maximum Counter setting (Default=type of piston) */
     int32_t distance;       /**< Difference between max and min counts */
-    float length;           /**< Length to calculate conversion factor */
-    float conversion_factor;    /**< Conversion factor to calculate "length" */
+    double length;           /**< Length to calculate conversion factor */
+    double conversion_factor;    /**< Conversion factor to calculate "length" */
 }sEncoderSettings_t;
 
 void ENC_Init(void);
 void ENC_FactoryReset(void);
-float ENC_Get_Length(void);
-void ENC_Set_MaxLength(float val) ;
+double ENC_Get_Length(void);
+void ENC_Set_MaxLength(double val) ;
 
 
 #ifdef TEST
@@ -59,7 +59,7 @@ void _set_max_enc_counter(void);
 bool _set_min_count(int32_t val);
 bool _set_max_count(int32_t val);
 void _calculate_encoder_distance(void);
-float _calculate_length(void);
+double _calculate_length(void);
 #endif
 
 #endif // _ENCODER_H
